Dropbox - extractor - update
refactoring #2
